move DailyFiguresDict
